TimeIndependentSchrodingerEquation.jl

Documentation of the TimeIndependentSchrodingerEquation library.

TimeIndependentSchrodingerEquation.eigen_values_and_eigen_vectors — Function

eigen_values_and_eigen_vectors(p,q,r,dΩ,USpace,VSpace; <keyword arguments>)

Compute matrix eigen problem and return eigenvalues and eigenvectors

...

Arguments

p,q,r::Function: functions from Sturm Liouville formulation

`dΩ::Gridap.CellData.GenericMeasure : measure of FE grid

USpace::FESpace: trial FE Space

VSpace::FESpace: test FE Space

`params::Tuple=(nev,tol,maxiter,explicittransform,sigma): params to Arpack eigs function.

nev::Integer=10: quantity of eigendata to calculate

tol::Float64=10e-9: relative tolerance for convergence of Ritz values

maxiter::Integer=100: maximum number of iterations

explicittransform::Symbol=:none: shift and invert should be explicitly invoked in julia code

sigma::Float64=1.0: the level shift used in inverse iteration.

...

TimeIndependentSchrodingerEquation.eigen_problem — Function

eigen_problem(weakformₖ,weakformₘ,test,trial;; <keyword arguments>)

Define eigen problem as an input to solve function where we compute eigen problem by Arpack eigs function.

...
# Arguments
- `weakformₖ::Function`: forma bilineal lado izquierdo de la formulación débil
- `weakformₘ::Function`: forma bilineal lado derecho de la formulación débil
- `test::FESpace`: espacio de prueba, puede ser MultiFieldFESpace
- `trial::FESpace`: espacio de solución, puede ser MultiFieldFESpace
- `nev::Int=10`: número de autovalores requeridos
- `tol::Float64=10e-6`: relative tolerance for convergence of Ritz values
- `maxiter::Integer=100`: maximum number of iterations
- `explicittransform::Symbol=:none`: shift and invert should be explicitly invoked in julia code
                           `=:auto`:
                           `=:shiftinvert`:
- `sigma::Float64=1.0`: the level shift used in inverse iteration.
- `which::Symbol=:LM`: eigenvalues of largest magnitude (default)
               `=:SM`: eigenvalues of smallest magnitude
               `=:LR`: eigenvalues of largest real part
               `=:SR`: eigenvalues of smallest real part
               `=:LI`: eigenvalues of largest imaginary part (nonsymmetric or complex matrix only)
               `=:SI`: eigenvalues of smallest imaginary part (nonsymmetric or complex matrix only)
               `=:BE`: compute half of the eigenvalues from each end of the spectrum, biased in favor of the high end. (real symmetric matrix only)
...

TimeIndependentSchrodingerEquation.solve — Function

solve(prob)

Compute eigen problem by Arpack eigs function and return eigenvalues and eigenvectors.

...
# Arguments
...
`prob::EigenProblem`: problem deinition

TimeIndependentSchrodingerEquation.space_coord — Function

space_coord(dom,Δr,n;<keyword arguments>)

Returns coordinate vector (r) and discrete points (pts) for 1D or 2D spaces. if dimension=="1D" ⇒ dom=(x₁,x₂); Δr=Δx; n=nx if dimension=="2d" ⇒ dom=(x₁,x₂,y₁,y₂); Δr=(Δx,Δy); n=(nx,ny) ...

Arguments

- `dom::Tuple`: FE cartesian domain.
- `Δr`:: discretization of FE space.
- `n`:: number of FE in each direction.

...

TimeIndependentSchrodingerEquation.fe_spaces — Function

fe_spaces(model,reff,grid_type; <keyword arguments>)

Create finite element (FE) spaces (Trial and Test spaces).

...

Arguments

BC_type::String="FullDirichlet": the type of boundary condition.

TypeData::Type=ComplexF64: the type of data to define FE spaces.

conf_type::Symbol=:H1: the regularity of the interpolation at the boundaries of cells in the mesh. (e.g.:L2,:H1,:C0,:Hgrad,)

...

Warning!

***->> The documentation is under construction <<-***