Documentation of the TimeIndependentSchrodingerEquation module.

TimeIndependentSchrodingerEquation.bilineal_forms — Method

bilineal_forms(p,q,r,dΩ)

Aim

Returns bilineals forms (a(u,v) and b(u,v)) for eigenvalues 1D or 2D (equal masses)

Arguments

p::Function: kinetic energy function from Sturm Liouville_Formulation.

q::Function: potential energy function from Sturm Liouville_Formulation.

r::Function: weight or density function from Sturm Liouville_Formulation.

dΩ::Gridap.CellData.GenericMeasure: differential FE domain

TimeIndependentSchrodingerEquation.default_solver_eigen_problem — Method

default_solver_eigen_problem(params)

Aim

Function to resolve unidimensonal eigen problem

Arguments

params::Params1D: parameters of 1D potential

TimeIndependentSchrodingerEquation.default_solver_eigen_problem — Method

default_solver_eigen_problem(params)

Aim

Function to resolve bidimensonal eigen problem over cartesian grid

Arguments

params::Params2D: parameters fo 2D potential

different_masses::Tuple: keyword to specify if we want to simulate two particles with diferent masses

TimeIndependentSchrodingerEquation.eigen_problem — Method

eigen_problem(weakformₖ,weakformₘ,test,trial;; <keyword arguments>)

Aim

Define eigen problem as an input to solve function where we compute eigen problem by Arpack eigs function.

Arguments

weakformₖ::Function: forma bilineal lado izquierdo de la formulación débil

weakformₘ::Function: forma bilineal lado derecho de la formulación débil

test::FESpace: espacio de prueba, puede ser MultiFieldFESpace

trial::FESpace: espacio de solución, puede ser MultiFieldFESpace

nev::Int=10: número de autovalores requeridos

tol::Float64=10e-6: relative tolerance for convergence of Ritz values

maxiter::Integer=100: maximum number of iterations

explicittransform::Symbol=:none: shift and invert should be explicitly invoked in julia code

=:auto:

=:shiftinvert:

sigma::Float64=1.0: the level shift used in inverse iteration.

which::Symbol=:LM: eigenvalues of largest magnitude (default)

=:SM: eigenvalues of smallest magnitude

=:LR: eigenvalues of largest real part

=:SR: eigenvalues of smallest real part

=:LI: eigenvalues of largest imaginary part (nonsymmetric or complex matrix only)

=:SI: eigenvalues of smallest imaginary part (nonsymmetric or complex matrix only)

=:BE: compute half of the eigenvalues from each end of the spectrum, biased in favor of the high end. (real symmetric matrix only)

TimeIndependentSchrodingerEquation.eigen_values_and_eigen_vectors — Method

eigen_values_and_eigen_vectors(p,q,r,dΩ,USpace,VSpace; <keyword arguments>)

Aim

Compute matrix eigen problem and return eigenvalues and eigenvectors

Arguments

p,q,r::Function: functions from Sturm Liouville formulation

dΩ::Gridap.CellData.GenericMeasure: measure of FE grid

USpace::FESpace: trial FE Space

VSpace::FESpace: test FE Space

params::Tuple=(nev,tol,maxiter,explicittransform,sigma): params to Arpack eigs function.

nev::Integer=10: quantity of eigendata to calculate

tol::Float64=10e-9: relative tolerance for convergence of Ritz values

maxiter::Integer=100: maximum number of iterations

explicittransform::Symbol=:none: shift and invert should be explicitly invoked in julia code

sigma::Float64=1.0: the level shift used in inverse iteration.

TimeIndependentSchrodingerEquation.fe_spaces — Method

fe_spaces(model,reff,grid_type; <keyword arguments>)

Aim

Create finite element (FE) spaces (Trial and Test spaces).

Arguments

BC_type::String="FullDirichlet": the type of boundary condition.

TypeData::Type=ComplexF64: the type of data to define FE spaces.

conf_type::Symbol=:H1: the regularity of the interpolation at the boundaries of cells in the mesh. (e.g.:L2,:H1,:C0,:Hgrad,)

TimeIndependentSchrodingerEquation.input_data — Method

input_data(data_file_name)

Aim

Definition of input data form input.dat file using specific type structures.

Arguments

data_file_name::String: name of input data file

TimeIndependentSchrodingerEquation.measures — Method

measures(model,degree,tags_boundary)

Aim

The triangulation and integration aproximated Lebesgue measures

Arguments

model: FE grid model.

degree::Integer: degree of quadrature rule to use in the cells of triangulation.

tags_boundary: tags of boundary conditions.

TimeIndependentSchrodingerEquation.norm_l2 — Method

norm_l2(𝜳,dΩ)

Aim

Compute de L2 norm for specific FE wave function

Arguments

𝜳::CellField: specific FE wave function

dΩ::Gridap.CellData.GenericMeasure: integration domain

TimeIndependentSchrodingerEquation.read_bin — Method

read_bin(fileName;<keyword arguments>)

Aim

Read binary file

Arguments

fileName::String: name of file data

matrix_data::Bool: optional boolean keyword to specify matrix or array data

c_dim::Int: optional column number of matrix data

TimeIndependentSchrodingerEquation.read_bin — Method

read_bin(io,::Type{T},n,matrix_data,c_dim)

Aim

Speeds up the read binary file

Arguments

io::IO: in/output variable

::Type{T}: data type

n::Int: total number of elements in array/matrix data

matrix_data::Bool: boolean keyword to specify matrix or array data

c_dim::Int: column number of matrix data

TimeIndependentSchrodingerEquation.solve — Method

solve(prob)

Aim

Compute eigen problem by Arpack eigs function and return eigenvalues and eigenvectors.

Arguments

prob::EigenProblem: problem deinition

TimeIndependentSchrodingerEquation.space_coord — Method

space_coord(dom,Δr,n;<keyword arguments>)

Aim

Returns coordinate vector (r) and discrete points (pts) for 1D or 2D spaces.

if dimension=="1D" ⇒ dom=(x₁,x₂); Δr=Δx; n=nx

if dimension=="2d" ⇒ dom=(x₁,x₂,y₁,y₂); Δr=(Δx,Δy); n=(nx,ny)

Arguments

dom::Tuple: FE cartesian domain.

Δr:: discretization of FE space.

n:: number of FE in each direction.

TimeIndependentSchrodingerEquation.write_bin — Method

write_bin(x,fileName;<keyword arguments>)

Aim

Write binary file from matrix data

Arguments

x::Matrix{T}: matrix data

fileName::String: name of file data

existing_file::Bool=false: optional boolean keyword to delete or not delete existing data

TimeIndependentSchrodingerEquation.write_bin — Method

write_bin(x,fileName;<keyword arguments>)

Aim

Write binary file from array data

Arguments

x::Array{T,1}: array/vector data

fileName::String: name of file data

existing_file::Bool=false: optional boolean keyword to delete or not delete existing data